refactor: move debug logic in separate file to reduce main
