+++ cliv2/cmd/cliv2/main.go
@@ -29,7 +29,6 @@ import (
 	"github.com/snyk/go-application-framework/pkg/auth"
 	"github.com/snyk/go-application-framework/pkg/configuration"
 	localworkflows "github.com/snyk/go-application-framework/pkg/local_workflows"
-	"github.com/snyk/go-application-framework/pkg/logging"
 	"github.com/snyk/go-application-framework/pkg/networking"
 	"github.com/snyk/go-application-framework/pkg/runtimeinfo"
 	"github.com/snyk/go-application-framework/pkg/utils"
@@ -69,52 +68,6 @@ const (
 	handleErrorShowHelp            HandleError = iota
 	handleErrorUnhandled           HandleError = iota
 )
-
-func getDebugLevel(config configuration.Configuration, logger *zerolog.Logger) zerolog.Level {
-	loglevel := zerolog.DebugLevel
-	if loglevelString := config.GetString("snyk_log_level"); loglevelString != "" {
-		var err error
-		loglevel, err = zerolog.ParseLevel(loglevelString)
-		if err == nil {
-			logger.Log().Msgf("Setting log level to %s", loglevelString)
-		} else {
-			logger.Log().Msgf("%v", err)
-			loglevel = zerolog.DebugLevel
-		}
-	}
-	return loglevel
-}
-
-func initDebugLogger(config configuration.Configuration) *zerolog.Logger {
-	debug := config.GetBool(configuration.DEBUG)
-	if !debug {
-		return &noopLogger
-	} else {
-		var consoleWriter = zerolog.ConsoleWriter{
-			Out:        os.Stderr,
-			TimeFormat: time.RFC3339,
-			NoColor:    true,
-			PartsOrder: []string{
-				zerolog.TimestampFieldName,
-				"ext",
-				"separator",
-				zerolog.CallerFieldName,
-				zerolog.MessageFieldName,
-			},
-			FieldsExclude: []string{"ext", "separator"},
-			FormatTimestamp: func(i interface{}) string {
-				t, _ := time.Parse(time.RFC3339, i.(string))
-				return strings.ToUpper(fmt.Sprintf("%s", t.UTC().Format(time.RFC3339)))
-			},
-		}
-
-		scrubLogger := logging.NewScrubbingWriter(zerolog.MultiLevelWriter(consoleWriter), logging.GetScrubDictFromConfig(config))
-		localLogger := zerolog.New(scrubLogger).With().Str("ext", "main").Str("separator", "-").Timestamp().Logger()
-		loglevel := getDebugLevel(config, &localLogger)
-		debugLogger := localLogger.Level(loglevel)
-		return &debugLogger
-	}
-}
 
 func main() {
 	errorCode := MainWithErrorCode()
